Factor out BulkTileOperation
